+++ portal-ejb/src/com/liferay/portlet/workflow/service/spring/WorkflowComponentService.java
@@ -43,7 +43,7 @@ public interface WorkflowComponentService {
 		throws com.liferay.portal.kernel.jbi.WorkflowComponentException, 
 			java.rmi.RemoteException;
 
-	public long getDefinitionsCount(long definitionId, java.lang.String name)
+	public int getDefinitionsCount(long definitionId, java.lang.String name)
 		throws com.liferay.portal.kernel.jbi.WorkflowComponentException, 
 			java.rmi.RemoteException;
 
@@ -60,7 +60,7 @@ public interface WorkflowComponentService {
 		throws com.liferay.portal.kernel.jbi.WorkflowComponentException, 
 			java.rmi.RemoteException;
 
-	public long getInstancesCount(long definitionId, long instanceId,
+	public int getInstancesCount(long definitionId, long instanceId,
 		java.lang.String workflowName, java.lang.String workflowVersion,
 		java.lang.String gtStartDate, java.lang.String ltStartDate,
 		java.lang.String gtEndDate, java.lang.String ltEndDate,
@@ -110,7 +110,7 @@ public interface WorkflowComponentService {
 		throws com.liferay.portal.kernel.jbi.WorkflowComponentException, 
 			java.rmi.RemoteException;
 
-	public long getUserTasksCount(long instanceId, java.lang.String taskName,
+	public int getUserTasksCount(long instanceId, java.lang.String taskName,
 		java.lang.String workflowName, java.lang.String assignedTo,
 		java.lang.String gtCreateDate, java.lang.String ltCreateDate,
 		java.lang.String gtStartDate, java.lang.String ltStartDate,
